requirements.txt
absl-py==0.11.0
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
depthai==2.13.3.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageai==2.1.5
importlib-metadata==2.0.0
imutils==0.5.4
isort==5.6.4
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
matplotlib==3.3.3
mccabe==0.6.1
mock==4.0.2
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.3.0
piexif==1.1.3
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.6.0
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.7.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
